refactor: views 레이어 SRP 정리 — 각 Screen 을 "데이터 준비 → 가드 → 조립" 에 집중
화면 파일에 섞여있던 서브 컴포넌트·도메인 상수·렌더 분기 내부 함수를 적절한 레이어로 분리. 각 Screen 이 자기 책임만 담당하도록 정리. 순 154 LOC 감소 (87 추가 / 241 삭제).

shared/ui:
- ScreenLayout 에 showBackButton prop 추가 (기본 true). 탭 루트(Mypage, Search)에서 back button 없이 재사용 가능

features/auth 신규 (views 간 중복 제거):
- AuthBrandHeader — "epigram" 로고 + subtitle (Login, SignUp 공용)
- AuthToggleLink — 하단 전환 링크 블록 (Login, SignUp 공용)
- LogoutButton — MypageScreen 인라인이었던 것을 이동. useLogout 을 쓰므로 auth feature 가 소유하는 게 맞음

features/epigram-create:
- EPIGRAM_CREATE_DEFAULT_VALUES 상수를 model/defaultValues.ts 로 이동. 도메인 기본값은 feature 가 소유

views 내부 분리:
- views/add-epigram/ui/WritingTips.tsx — AddEpigramScreen 인라인 추출
- views/epigram-detail/ui/EpigramDetailActionMenu.tsx — 화면 내부 ActionMenu 분리 + useEpigramDelete 직접 호출 제거, onDelete/isDeleting prop 으로 decouple
- views/epigram-detail/ui/EpigramDetailBody.tsx — renderBody 내부 함수 제거
- views/epigram-edit/ui/EpigramEditBody.tsx — renderBody 내부 함수 제거
- views/search/ui/SearchInitialState.tsx — InitialState 인라인 추출

views 화면 정리:
- AddEpigramScreen: WritingTips·DEFAULT_VALUES 분리, 45 → 37 LOC
- EpigramDetailScreen: inline ActionMenu·renderBody 제거, 87 → ~35 LOC
- EpigramEditScreen: EpigramEditBody 분리 + 가드 순서 수정 (epigram 로드 후 소유권 체크하도록 reorder — 로딩 중에 redirect 되던 엣지 케이스 방지)
- LoginScreen, SignUpScreen: 공통 AuthBrandHeader + AuthToggleLink 로 대체
- MypageScreen: SafeAreaView 직접 사용 → ScreenLayout showBackButton={false}, LogoutButton import
- SearchScreen: SafeAreaView → ScreenLayout, SearchInitialState 분리

tsc / expo lint 통과. Alert 중복 정책(ActionMenu 메뉴 + useEpigramDelete 확인 2단계)은 현재 유지.
