+++ src/features/auth/index.ts
@@ -5,7 +5,10 @@ export type { SignUpFormValues } from "./model/signUpSchema";
 export { useLogout } from "./model/useLogout";
 export { useSessionExpiryRedirect } from "./model/useSessionExpiryRedirect";
 
+export { AuthBrandHeader } from "./ui/AuthBrandHeader";
 export { AuthGate } from "./ui/AuthGate";
+export { AuthToggleLink } from "./ui/AuthToggleLink";
 export { GuestLoginButton } from "./ui/GuestLoginButton";
 export { LoginForm } from "./ui/LoginForm";
+export { LogoutButton } from "./ui/LogoutButton";
 export { SignUpForm } from "./ui/SignUpForm";
