+++ src/views/login/ui/LoginScreen.tsx
@@ -1,15 +1,16 @@
 import { router } from "expo-router";
 import type { ReactElement } from "react";
-import { Pressable, ScrollView, Text, View } from "react-native";
+import { ScrollView, View } from "react-native";
 
-import { GuestLoginButton, LoginForm } from "~/features/auth";
+import {
+  AuthBrandHeader,
+  AuthToggleLink,
+  GuestLoginButton,
+  LoginForm,
+} from "~/features/auth";
 import { ScreenLayout } from "~/shared/ui";
 
 export function LoginScreen(): ReactElement {
-  function handleGoToSignUp(): void {
-    router.push("/signup");
-  }
-
   return (
     <ScreenLayout>
       <ScrollView
@@ -17,28 +18,16 @@ export function LoginScreen(): ReactElement {
         contentContainerClassName="flex-grow justify-center px-screen-x py-10"
         keyboardShouldPersistTaps="handled"
       >
-        <View className="mb-10 items-center gap-2">
-          <Text className="font-serif-bold text-4xl text-blue-800">
-            epigram
-          </Text>
-          <Text className="font-sans text-sm text-black-300">
-            인생 명언을 모아보세요
-          </Text>
-        </View>
+        <AuthBrandHeader subtitle="인생 명언을 모아보세요" />
         <LoginForm />
         <View className="mt-3">
           <GuestLoginButton />
         </View>
-        <View className="mt-6 flex-row justify-center gap-1">
-          <Text className="font-sans text-sm text-black-300">
-            아직 회원이 아니신가요?
-          </Text>
-          <Pressable onPress={handleGoToSignUp} accessibilityRole="link">
-            <Text className="font-sans text-sm font-semibold text-blue-600">
-              회원가입
-            </Text>
-          </Pressable>
-        </View>
+        <AuthToggleLink
+          prompt="아직 회원이 아니신가요?"
+          linkLabel="회원가입"
+          onPress={() => router.push("/signup")}
+        />
       </ScrollView>
     </ScreenLayout>
   );
